Share System Nexus span names
